Reaching the database
- successfully retrieving the user object with updated trips ids
