+++ react-client/src/components/CreateView/chooser/saveBox.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import $ from 'jquery';
 
 import RaisedButton from 'material-ui/RaisedButton';
 import FlatButton from 'material-ui/FlatButton';
@@ -15,19 +16,38 @@ class SaveBox extends React.Component {
     };
     this.updateTripName = this.updateTripName.bind(this);
     this.updateTripDesc = this.updateTripDesc.bind(this);
-    this.saveTrip = this.saveTrip.bind(this);
+    this.saveTriptoDB = this.saveTriptoDB.bind(this);
   }
 
-  saveTrip() {
-    var tripItem = {
-      trip: this.props.trip,
-      name: this.state.name,
-      description: this.state.description,
-      stepIndex: 1,
-      edit: true,
-    }
-    console.log('saved: ', tripItem);
-    this.props.changePage('LIST');
+  randomId() {
+    var a = Math.random(0, 1) * 1000000000
+    var b = a.toString();
+    var c = parseInt(b);
+    return c;
+  }
+
+  saveTriptoDB() {
+    let postData = {
+      id: this.randomId(),
+      food: this.props.trip.restaurants,
+      attractions: this.props.trip.attractions,
+      lodging: this.props.trip.hotels,
+      destination: this.props.destination,
+      facebookId: this.props.user.id
+    };
+    console.log(postData);
+    $.ajax({
+      url: '/save',
+      method: 'POST',
+      data: JSON.stringify(postData),
+      contentType: 'application/json',
+      success: (data) => {
+        console.log('the POST request went through! ', data);
+      },
+      error: (err) => {
+        console.log('err', err);
+      }
+    });
   }
 
   updateTripName(event) {
@@ -53,7 +73,7 @@ class SaveBox extends React.Component {
       <RaisedButton
         label="Save"
         primary={true}
-        onClick={this.saveTrip}
+        onClick={this.saveTriptoDB}
       />
     ];
